+++ simtools.cc
@@ -171,7 +171,7 @@ void init_perlin_map( sint32 w, sint32 h )
 	map = new float[map_w*(h+2)];
 	for(  sint32 y=0;  y<h+2;  y++ ) {
 		for(  sint32 x=0;  x<map_w;  x++ ) {
-			map[x+(y*map_w)] = int_noise( x-1, y-1 );
+			map[x+(y*map_w)] = (float)int_noise( x-1, y-1 );
 		}
 	}
 }
